backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
requests==2.28.2
python-dotenv==1.0.0
django-cors-headers==4.1.0
dj-database-url==2.0.0
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0 